put the timer and trigger traits in two files
